+++ src/Player.java
@@ -11,13 +11,13 @@ abstract class Player {
      * An integer to uniquely identify each player. Since this implementation only considers the two-player version
      * of Carcassonne, the ID should either be 1 or 2.
      */
-    protected byte playerID;
-    protected Player(byte playerID) {
+    protected int playerID;
+    protected Player(int playerID) {
         if (playerID != 1 && playerID != 2) {
             //logger.error("Player initialised with invalid ID");
         }
         this.playerID = playerID;
     }
 
-    abstract byte[] decideOnNextMove(GameState state, GameStateSpace stateSpace, Tile tile, List<Tile> deck) throws Exception;
+    abstract int[] decideOnNextMove(GameState state, GameStateSpace stateSpace, Tile tile, List<Tile> deck) throws Exception;
 }
