+++ src/RandomPlayer.java
@@ -10,7 +10,7 @@ public class RandomPlayer extends Player {
     Random random;
 
     public RandomPlayer(int playerID) {
-        super((byte) playerID);
+        super((int) playerID);
         random = new Random();
     }
 
@@ -19,7 +19,7 @@ public class RandomPlayer extends Player {
      * @return The coordinates of where the tile will be placed.
      */
     @Override
-    public byte[] decideOnNextMove(GameState state, GameStateSpace stateSpace, Tile tile, List<Tile> deck) {
+    public int[] decideOnNextMove(GameState state, GameStateSpace stateSpace, Tile tile, List<Tile> deck) {
 
         //GameState state = new GameState(originalState);
 
@@ -27,7 +27,7 @@ public class RandomPlayer extends Player {
 
         // In this case no move is possible and a new tile has to be drawn.
         if (ars.isEmpty()) {
-            return new byte[]{-1, -1};
+            return new int[]{-1, -1};
         }
 
         ActionRotationPair move = ars.get(random.nextInt(ars.size()));
@@ -36,13 +36,13 @@ public class RandomPlayer extends Player {
             tile.rotate();
         }
 
-        byte[] action = move.getAction();
+        int[] action = move.getAction();
 
         //logger.info("Player {} placed tile at [{}, {}] with rotation {}.", playerID, move.getAction()[0], move.getAction()[1], tile.getRotation());
 
         // Here the placement of the meeple is decided
 
-        List<Byte> meeplePlacements = stateSpace.legalMeeples(state, tile, action);
+        List<Integer> meeplePlacements = stateSpace.legalMeeples(state, tile, action);
 
         if (random.nextBoolean() && !meeplePlacements.isEmpty() && state.getNumMeeples(playerID) > 0) {
             //logger.info("Player {} places meeple on point {}, which is of type {} and belongs to area {}. {} meeples remaining.", tile.getMeeple()[1], tile.getMeeple()[0], tile.getPoint(tile.getMeeple()[0]), tile.getArea(tile.getMeeple()[0]), state.getNumMeeples(playerID));
